requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-Cors==4.0.0
flask-swagger-ui==4.11.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mercadopago==2.2.1
mysql-connector-python==8.3.0
parse==1.20.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
ratelimit==2.2.1
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.1
